+++ integrationtests/src/jvmTest/kotlin/de/connect2x/trixnity/messenger/integrationtests/util/TestUtils.kt
@@ -5,11 +5,10 @@ import de.connect2x.lognity.api.logger.Level
 import de.connect2x.lognity.test.TestBackend
 
 fun configureTestLogging() {
-    Backend.setOnce(TestBackend) {
-        val defaultConfig = TestBackend.configSpec
-        TestBackend.configSpec = {
-            defaultConfig()
-            level = Level.INFO // DockerJava produces too much noise
-        }
+    Backend.set(TestBackend)
+    val defaultConfig = TestBackend.configSpec
+    TestBackend.configSpec = {
+        defaultConfig()
+        level = Level.INFO // DockerJava produces too much noise
     }
 }