+++ trixnity-messenger/src/jsMain/kotlin/de/connect2x/trixnity/messenger/JsJodaTimeZoneModule.kt
@@ -1,7 +1,0 @@
-package de.connect2x.trixnity.messenger
-
-@JsModule("@js-joda/timezone")
-@JsNonModule
-external object JsJodaTimeZoneModule
-
-private val jsJodaTz = JsJodaTimeZoneModule